Make lambda use the ls repo

+++ 2-build-layer.sh
@@ -1,11 +1,29 @@
 #!/bin/bash
 
-# Copy only the desired files over from the compiler /TODO
-# Clone repo
-#cp lattice-surgery-compiler/*.py function
-#cp lattice-surgery-compiler/requirements.txt function
 
 set -eo pipefail
-rm -rf package
-cd function
-pip install --target ../package/python -r requirements.txt
+
+echo "Updating submodules"
+git submodule update --remote
+echo "Moving the files from the handler and the lattice-surgery-compiler into a single folder"
+rm -rf function
+mkdir function
+cp -r handler/* function
+rm function/requirements.txt
+cp -r lattice-surgery-compiler/* function
+rm function/requirements.txt
+rm -r function/web
+rm -r function/webgui
+
+
+if [ package -ot handler/requirements.txt ] || [ package -ot lattice-surgery-compiler/requirements.txt ]; then
+    echo "Creating the layer. This might take some time"
+    merge_requirements  lattice-surgery-compiler/requirements.txt handler/requirements.txt
+    mv requirements-merged.txt function/requirements.txt
+    rm -rf package
+    cd function
+    pip install --target ../package/python -r requirements.txt
+else
+    echo "Using existing layer"
+fi
+
